+++ examples/SimpleClient.hs
@@ -9,7 +9,7 @@ import Data.Text.IO as TIO
 main :: IO ()
 main = do
   -- 1. create a ConnectionInfo
-  let connInfo = RemoteConnectionInfo "mytestdb" "127.0.0.1" (show defaultServerPort) emptyNotificationCallback
+  let connInfo = RemoteConnectionInfo "mytestdb" defaultRemoteServerAddress emptyNotificationCallback
   -- 2. connected to the remote database
   eConn <- connectProjectM36 connInfo
   case eConn of