+++ src/bin/ProjectM36/Server/RemoteCallTypes/Json.hs
@@ -182,6 +182,9 @@ instance FromJSON Attributes where
 instance ToJSON RelationalError
 instance FromJSON RelationalError
 
+instance ToJSON NotificationExpression
+instance FromJSON NotificationExpression
+
 instance ToJSON SQLError
 instance FromJSON SQLError