+++ src/bin/TutorialD/Interpreter/RODatabaseContextOperator.hs
@@ -21,6 +21,7 @@ import Data.Maybe
 import Data.Functor
 
 --operators which only rely on database context reading
+
 data RODatabaseContextOperator =
   ShowRelation RelationalExpr |
   PlotRelation RelationalExpr |
@@ -32,6 +33,7 @@ data RODatabaseContextOperator =
   ShowRelationVariables |
   ShowAtomFunctions |
   ShowDatabaseContextFunctions |
+  ShowNotifications |
   ShowDataFrame DF.DataFrameExpr |
   GetDDLHash |
   ShowDDL |
@@ -71,6 +73,9 @@ showAtomFunctionsP = colonOp ":showatomfunctions" >> pure ShowAtomFunctions
 showDatabaseContextFunctionsP :: Parser RODatabaseContextOperator
 showDatabaseContextFunctionsP = colonOp ":showdatabasecontextfunctions" >> pure ShowDatabaseContextFunctions
 
+showNotificationsP :: Parser RODatabaseContextOperator
+showNotificationsP = colonOp ":shownotifications" >> pure ShowNotifications
+
 quitP :: Parser RODatabaseContextOperator
 quitP = do
   colonOp ":quit"
@@ -97,6 +102,7 @@ roDatabaseContextOperatorP = typeP
              <|> showTypesP
              <|> showAtomFunctionsP
              <|> showDatabaseContextFunctionsP
+             <|> showNotificationsP
              <|> showDataFrameP
              <|> ddlHashP
              <|> showDDLP
@@ -171,6 +177,12 @@ evalRODatabaseContextOp sessionId conn ShowAtomFunctions = do
     
 evalRODatabaseContextOp sessionId conn ShowDatabaseContextFunctions = do
   eRel <- C.databaseContextFunctionsAsRelation sessionId conn
+  case eRel of
+    Left err -> pure $ DisplayErrorResult (T.pack (show err))
+    Right rel -> evalRODatabaseContextOp sessionId conn (ShowRelation (ExistingRelation rel))
+    
+evalRODatabaseContextOp sessionId conn ShowNotifications = do
+  eRel <- C.notificationsAsRelation sessionId conn
   case eRel of
     Left err -> pure $ DisplayErrorResult (T.pack (show err))
     Right rel -> evalRODatabaseContextOp sessionId conn (ShowRelation (ExistingRelation rel))