+++ src/lib/ProjectM36/Server.hs
@@ -120,7 +120,10 @@ requestHandlers testFlag ti =
                         handleConvertSQLQuery ti sessionId conn q),
      RequestHandler (\sState (ConvertSQLUpdates sessionId updates) -> do
                         conn <- getConn sState
-                        handleConvertSQLUpdates ti sessionId conn updates)
+                        handleConvertSQLUpdates ti sessionId conn updates),
+     RequestHandler (\sState (RetrieveNotificationsAsRelation sessionId) -> do
+                        conn <- getConn sState
+                        handleRetrieveNotificationsAsRelation ti sessionId conn)
      ] ++ if testFlag then testModeHandlers ti else []
 
 getConn :: ConnectionState ServerState -> IO Connection
@@ -200,6 +203,7 @@ initialServerState dbName conn =
   clientMap <- StmMap.new
   StmMap.insert conn dbName dbmap
   pure (ServerState { stateDBMap = dbmap, stateClientMap = clientMap })
+
 -- | A synchronous function to start the project-m36 daemon given an appropriate 'ServerConfig'. Note that this function only returns if the server exits. Returns False if the daemon exited due to an error. If the second argument is not Nothing, the port is put after the server is ready to service the port.
 launchServer :: ServerConfig -> Maybe (MVar SockAddr) -> IO Bool
 launchServer daemonConfig mAddr = do
@@ -214,22 +218,12 @@ launchServer daemonConfig mAddr = do
           hPutStrLn stderr ("Failed to create database connection: " ++ show err)
           pure False
         Right conn -> do
-          let hostname = bindHost daemonConfig
-              port = fromIntegral (bindPort daemonConfig)
+          let mTimeout = fromIntegral <$>
+                case perRequestTimeout daemonConfig of
+                  0 -> Nothing
+                  v -> Just v
+          (sockSpec, sockAddr) <- resolveRemoteServerAddress (bindAddress daemonConfig)
+          sState <- initialServerState (databaseName daemonConfig) conn
+          serve (requestHandlers (testMode daemonConfig) mTimeout) sState sockSpec sockAddr mAddr
 
-
-          --curryer only supports IPv4 for now
-          let addrHints = defaultHints { addrSocketType = Stream, addrFamily = AF_INET }
-          hostAddrs <- getAddrInfo (Just addrHints) (Just hostname) Nothing
-          case hostAddrs of
-            [] -> hPutStrLn stderr ("Failed to resolve: " <> hostname) >> pure False
-            (AddrInfo _ _ _ _ (SockAddrInet _ addr32) _):_ -> do
-              let hostAddr = hostAddressToTuple addr32
-                  mTimeout = fromIntegral <$> case perRequestTimeout daemonConfig of
-                                              0 -> Nothing
-                                              v -> Just v
-                  
-              sState <- initialServerState (databaseName daemonConfig) conn
-              serve (requestHandlers (testMode daemonConfig) mTimeout) sState hostAddr port mAddr
-            _ -> error "unsupported socket addressing mode (IPv4 only currently)"